% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celsify_temp.R
\name{celsify_temp}
\alias{celsify_temp}
\title{Convert the \code{temp} column from Farenheit to Celsius}
\usage{
celsify_temp(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A data frame.
}
\description{
Convert the \code{temp} column from Farenheit to Celsius
}
\examples{
library(tibble)

data <- tibble(temp = c(32, 0), english = c("US", "UK"))
celsify_temp(data)
}
