% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localize_beach.R
\name{localize_beach}
\alias{localize_beach}
\title{Localize a place depending on the word used to describe it}
\usage{
localize_beach(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A data frame.
}
\description{
Localize a place depending on the word used to describe it
}
\examples{
library(tibble)

data <- tibble(where = c("beach", "seashore"))
localize_beach(data)
}
